+++ tools/openapi-templates/go/api.mustache
@@ -8,7 +8,7 @@ import (
 	"io"
 	"net/http"
 	"net/url"
-    "sensorbucket.nl/sensorbucket/pkg/web"
+    "sensorbucket.nl/sensorbucket/internal/web"
 {{#imports}}	"{{import}}"
 {{/imports}}
 )
